door & window -> openLock
